rename to no-internal-modules
