+++ src/rules/no-internal-modules.js
