+++ test-build.js
@@ -1,2 +1,2 @@
 require('babel-register')
-require('./tests/src/rules/no-reaching-inside.js')
+require('./tests/src/rules/no-internal-modules.js')
